+++ src/components/CheckForm/FormLayouts/CheckTracerouteLayout.tsx
@@ -1,7 +1,4 @@
 import React from 'react';
-import { GrafanaTheme2 } from '@grafana/data';
-import { useStyles2 } from '@grafana/ui';
-import { css } from '@emotion/css';
 
 import { CheckFormValuesTraceroute, CheckType } from 'types';
 import { CheckEnabled } from 'components/CheckEditor/FormComponents/CheckEnabled';
@@ -18,11 +15,12 @@ import { CheckUsage } from 'components/CheckUsage';
 import { LabelField } from 'components/LabelField';
 
 export const CheckTracerouteLayout = () => {
-  const styles = useStyles2(getStyles);
-
   return (
     <FormLayout>
-      <FormLayout.Section label="General settings">
+      <FormLayout.Section
+        label="General settings"
+        fields={[`enabled`, `job`, `target`, `probes`, `frequency`, `timeout`]}
+      >
         <CheckEnabled />
         <CheckJobName />
         <CheckTarget checkType={CheckType.Traceroute} />
@@ -30,23 +28,23 @@ export const CheckTracerouteLayout = () => {
         <CheckPublishedAdvanceMetrics />
         <CheckUsage />
       </FormLayout.Section>
-      <FormLayout.Section label="Advanced options">
-        <div className={styles.maxWidth}>
-          <LabelField<CheckFormValuesTraceroute> />
-          <TracerouteMaxHops />
-          <TracerouteMaxUnknownHops />
-          <TraceroutePTRLookup />
-        </div>
+      <FormLayout.Section
+        label="Advanced options"
+        fields={[
+          `labels`,
+          `settings.traceroute.maxHops`,
+          `settings.traceroute.maxUnknownHops`,
+          `settings.traceroute.ptrLookup`,
+        ]}
+      >
+        <LabelField<CheckFormValuesTraceroute> />
+        <TracerouteMaxHops />
+        <TracerouteMaxUnknownHops />
+        <TraceroutePTRLookup />
       </FormLayout.Section>
-      <FormLayout.Section label="Alerting">
+      <FormLayout.Section label="Alerting" fields={[`alertSensitivity`]}>
         <CheckFormAlert />
       </FormLayout.Section>
     </FormLayout>
   );
 };
-
-const getStyles = (theme: GrafanaTheme2) => ({
-  maxWidth: css({
-    maxWidth: `500px`,
-  }),
-});