+++ src/scenes/Common/avgLatencyStat.ts
@@ -3,24 +3,38 @@ import { DataSourceRef, ThresholdsMode } from '@grafana/schema';
 
 import { ExplorablePanel } from 'scenes/ExplorablePanel';
 
-function getQueryRunner(metrics: DataSourceRef) {
-  return new SceneQueryRunner({
+import { divideSumByCountTransformation } from './divideSumByCountTransformation';
+
+function getQueryRunner(metrics: DataSourceRef, minStep: string) {
+  const runner = new SceneQueryRunner({
     datasource: metrics,
+    minInterval: minStep,
+    maxDataPoints: 10,
     queries: [
       {
-        expr: 'sum(rate(probe_all_duration_seconds_sum{probe=~"$probe", instance="$instance", job="$job"}[$__range])) / sum(rate(probe_all_duration_seconds_count{probe=~"$probe", instance="$instance", job="$job"}[$__range]))',
+        expr: 'sum(rate(probe_all_duration_seconds_sum{probe=~"$probe", instance="$instance", job="$job"}[$__rate_interval]))',
         hide: false,
-        instant: true,
-        interval: '',
-        legendFormat: '',
+        instant: false,
+        range: true,
+        legendFormat: 'sum',
+        refId: 'A',
+      },
+      {
+        expr: 'sum(rate(probe_all_duration_seconds_count{probe=~"$probe", instance="$instance", job="$job"}[$__rate_interval]))',
+        hide: false,
+        instant: false,
+        range: true,
+        legendFormat: 'count',
         refId: 'B',
       },
     ],
   });
+
+  return divideSumByCountTransformation(runner);
 }
 
-export function getAvgLatencyStat(metrics: DataSourceRef) {
-  const queryRunner = getQueryRunner(metrics);
+export function getAvgLatencyStat(metrics: DataSourceRef, minStep: string) {
+  const queryRunner = getQueryRunner(metrics, minStep);
   return new ExplorablePanel({
     pluginId: 'stat',
     title: 'Average latency',