+++ src/scenes/Common/uptimeStat.ts
@@ -1,32 +1,54 @@
-import { SceneQueryRunner } from '@grafana/scenes';
+import { SceneDataTransformer, SceneQueryRunner } from '@grafana/scenes';
 import { DataSourceRef, ThresholdsMode } from '@grafana/schema';
 
 import { UPTIME_DESCRIPTION } from 'components/constants';
 import { ExplorablePanel } from 'scenes/ExplorablePanel';
 
-function getQueryRunner(metrics: DataSourceRef) {
-  return new SceneQueryRunner({
+function getQueryRunner(metrics: DataSourceRef, minStep: string) {
+  const runner = new SceneQueryRunner({
     datasource: metrics,
     queries: [
       {
+        editorMode: 'code',
         exemplar: true,
-        expr: '# find the average uptime over the entire time range evaluating \'up\' in 5 minute windows\navg_over_time(\n  (\n    # the inner query is going to produce a non-zero value if there was at least one successful check during the 5 minute window\n    # so make it a 1 if there was at least one success and a 0 otherwise\n    ceil(\n      # the number of successes across all probes\n      sum by (instance, job) (increase(probe_all_success_sum{instance="$instance", job="$job"}[5m]))\n      /\n      # the total number of times we checked across all probes\n      (sum by (instance, job) (increase(probe_all_success_count{instance="$instance", job="$job"}[5m])) + 1) # + 1 because we want to make sure it goes to 1, not 2\n    )\n  )\n  [$__range:5m]\n)',
+        expr: `# the inner query is going to produce a non-zero value if there was at least one successful check during the 5 minute window
+        # so make it a 1 if there was at least one success and a 0 otherwise
+        ceil(
+          # the number of successes across all probes
+          sum by (instance, job) (increase(probe_all_success_sum{instance="$instance", job="$job"}[$__rate_interval]))
+          /
+          # the total number of times we checked across all probes
+          (sum by (instance, job) (increase(probe_all_success_count{instance="$instance", job="$job"}[$__rate_interval])) + 1) # + 1 because we want to make sure it goes to 1, not 2
+        )`,
         hide: false,
-        instant: true,
-        interval: '',
+        instant: false,
+        interval: minStep,
         legendFormat: '',
-        refId: 'uptimeStat',
+        range: true,
+        refId: 'B',
+      },
+    ],
+  });
+
+  return new SceneDataTransformer({
+    $data: runner,
+    transformations: [
+      {
+        id: 'reduce',
+        options: {
+          reducers: ['mean'],
+        },
       },
     ],
   });
 }
 
-export function getUptimeStat(metrics: DataSourceRef) {
+export function getUptimeStat(metrics: DataSourceRef, minStep: string) {
   return new ExplorablePanel({
     pluginId: 'stat',
     title: 'Uptime',
     description: UPTIME_DESCRIPTION,
-    $data: getQueryRunner(metrics),
+    $data: getQueryRunner(metrics, minStep),
     fieldConfig: {
       defaults: {
         decimals: 2,