+++ src/scenes/Summary/latencyTimeseries.ts
@@ -11,7 +11,7 @@ function getLatencyQueryRunner(metrics: DataSourceRef) {
         expr: `
         (
           sum(
-            rate(probe_all_duration_seconds_sum{probe=~"$probe"}[$__range])
+            rate(probe_all_duration_seconds_sum{probe=~"$probe"}[$__rate_interval])
             * on (instance, job, probe, config_version) group_left max(sm_check_info{check_name=~"$check_type", region=~"$region", $Filters})
             by (instance, job, probe, config_version)
           )
@@ -20,7 +20,7 @@ function getLatencyQueryRunner(metrics: DataSourceRef) {
         /
         (
           sum(
-            rate(probe_all_duration_seconds_count{probe=~"$probe"}[$__range])
+            rate(probe_all_duration_seconds_count{probe=~"$probe"}[$__rate_interval])
             * on (instance, job, probe, config_version) group_left max(sm_check_info{check_name=~"$check_type", region=~"$region", $Filters})
             by (instance, job, probe, config_version)
           )
@@ -28,12 +28,11 @@ function getLatencyQueryRunner(metrics: DataSourceRef) {
         )
         `,
         hide: false,
-        interval: '',
+        interval: '1m',
         legendFormat: '{{job}}/{{ instance }}',
         refId: 'A',
       },
     ],
-    // maxDataPoints: 100,
   });
   return queryRunner;
 }
@@ -52,7 +51,7 @@ export function getLatencyTimeseriesPanel(metrics: DataSourceRef) {
           lineWidth: 2,
           fillOpacity: 0,
           gradientMode: 'none',
-          spanNulls: false,
+          spanNulls: true,
           insertNulls: false,
           showPoints: 'never',
           pointSize: 5,